+++ client/src/pages/home/home.jsx
@@ -25,7 +25,7 @@ const Home = () => {
         <EmojiEmotionsIcon />,
         <FavoriteIcon />,
         // <MedicationIcon />,
-        // <AppleIcon />,
+        <AppleIcon />,
         <BedIcon />,
         <FitnessCenterIcon />,
     ];
@@ -33,8 +33,8 @@ const Home = () => {
     // const item_list = ['Activity', 'Body Measurements', 'Moods', 'Heart', 'Medications', 'Nutrition', 'Sleep', 'Fitness']
     // const linklist = ['/activity', '/body-measurements', 'moods', 'heart', 'medications', 'nutrition', 'sleep', 'fitness']
 
-    const item_list = ['Activity', 'Body Measurements', 'Moods', 'Heart', 'Sleep', 'Fitness']
-    const linklist = ['/activity', '/body-measurements', 'moods', 'heart', 'sleep', 'fitness']
+    const item_list = ['Activity', 'Body Measurements', 'Moods', 'Heart', 'Nutrition', 'Sleep', 'Fitness']
+    const linklist = ['/activity', '/body-measurements', '/moods', '/heart', '/nutrition', '/sleep', '/fitness']
 
     const item_summaries = [
         'Track your daily physical activities, steps, and movement patterns to maintain an active lifestyle.',
@@ -42,7 +42,7 @@ const Home = () => {
         'Log your emotional states and mood changes to better understand your mental wellbeing over time.',
         'Keep track of heart-related data such as heart rate and blood pressure to monitor cardiovascular health.',
         // 'Manage your medications schedule, dosage, and reminders to ensure proper adherence.',
-        // 'Track your meals, calories, and nutrient intake to support a balanced and healthy diet.',
+        'Track your meals, calories, and nutrient intake to support a balanced and healthy diet.',
         'Record sleep duration and quality to improve rest and overall wellness.',
         'Log workouts, exercise routines, and fitness goals to stay motivated and track performance improvements.'
     ];
