+++ server/api/urls.py
@@ -16,7 +16,8 @@ urlpatterns = [
     path("activity/", include("activity.urls")),
     path("weightlog/", include("weightlog.urls")),
     path("heartRate/", include("heartRate.urls")),
-    path("sleep/", include("sleepRecord.urls")),
+    path("sleep/", include("sleep.urls")),
+    path("nutrition/", include("nutrition.urls")),
     path("mood/", include("mood.urls")),
     path("fitness/", include("fitnessRecord.urls")),
