+++ server/website/settings.py
@@ -64,7 +64,8 @@ INSTALLED_APPS = [
     "userInfo",
     "fitnessRecord",
     "heartRate",
-    "sleepRecord",
+    "sleep",
+    "nutrition",
     "activity",
     "weightlog",
     "mood",
